+++ src/main/ErrorHandler/errorHandler.interface.ts
@@ -1,0 +1,6 @@
+interface IErrorHandler {
+  handleError: HandleErrorFunction;
+  setCustomErrorHandler: SetCustomErrorHandlerFunction;
+}
+
+export default IErrorHandler;