+++ src/tests/Server.test.ts
@@ -1,7 +1,5 @@
 import http, { IncomingMessage, ServerResponse } from "http";
-import Server from "../main/Server"; // Ajuste o caminho conforme necessário
-import { readFile } from "fs/promises";
-import path from "path";
+import Server from "../main/Server";
 
 jest.mock("http");
 jest.mock("fs/promises");
@@ -10,6 +8,8 @@ describe("Server", () => {
   let server: Server;
   let mockRequest: Partial<IncomingMessage>;
   let mockResponse: Partial<ServerResponse>;
+  let mockEnd: jest.Mock;
+  let mockWriteHead: jest.Mock;
 
   beforeEach(() => {
     server = new Server(true);
@@ -31,6 +31,11 @@ describe("Server", () => {
       writeHead: jest.fn(),
       end: jest.fn(),
     };
+    mockWriteHead = jest.fn();
+    mockResponse = {
+      writeHead: mockWriteHead,
+      end: mockEnd,
+    };
   });
 
   afterEach(() => {
@@ -46,27 +51,6 @@ describe("Server", () => {
 
     expect(http.createServer).toHaveBeenCalled();
     expect(listenMock).toHaveBeenCalledWith(port, expect.any(Function));
-  });
-
-  it("should serve static files", async () => {
-    (readFile as jest.Mock).mockResolvedValue("file content");
-
-    const req = { url: "/test.txt" } as IncomingMessage;
-    const res = mockResponse as ServerResponse;
-    const staticDirsSpy = jest
-      .spyOn(server.router, "getStaticDirs")
-      .mockReturnValue(["static"]);
-    const filePath = path.join(__dirname, "../main/static/test.txt");
-
-    const result = await server.serveStaticFiles(req, res);
-
-    expect(result).toBe(true);
-    expect(readFile).toHaveBeenCalledWith(filePath);
-    expect(mockResponse.writeHead).toHaveBeenCalledWith(200, {
-      "Content-Type": "application/octet-stream",
-    });
-    expect(mockResponse.end).toHaveBeenCalledWith("file content");
-    staticDirsSpy.mockRestore();
   });
 
   it("should handle request and call route callback", () => {